server/requirements.txt
ago==0.0.93
attrs==19.3.0
Automat==0.8.0
awscli==1.16.287
beautifulsoup4==4.8.1
blinker==1.4
botocore==1.13.23
bs4==0.0.1
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
colorama==0.4.1
constantly==15.1.0
cryptography==2.8
cssselect==1.1.0
dnspython==1.16.0
docutils==0.15.2
dotmap==1.3.8
elasticsearch==7.1.0
feedfinder2==0.0.4
feedparser==5.2.1
firebase-admin==4.0.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Mail==0.9.1
google-api-core==1.16.0
google-api-python-client==1.8.0
google-auth==1.13.1
google-auth-httplib2==0.0.3
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
hjson==3.0.1
httplib2==0.17.1
hurry.filesize==0.9
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.10.1
jmespath==0.9.4
langdetect==1.0.7
lxml==4.4.1
MarkupSafe==1.1.1
msgpack==1.0.0
news-please==1.4.24
newsapi-python==0.2.5
newspaper3k==0.2.8
nltk==3.4.5
numpy==1.17.2
pandas==0.25.1
parsel==1.5.2
pathlib==1.0.1
Pillow==6.2.1
plac==1.1.3
Protego==0.1.15
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyDispatcher==2.0.5
PyHamcrest==1.9.0
pymongo==3.9.0
PyMySQL==0.9.3
pyOpenSSL==19.1.0
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.2
pywin32==227
PyYAML==5.1.2
queuelib==1.5.0
readability-lxml==0.7.1
redis==3.4.1
requests==2.21.0
requests-file==1.4.3
rq==1.3.0
rsa==3.4.2
s3transfer==0.2.1
Scrapy==1.8.0
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.5
textblob==0.15.3
tinysegmenter==0.3
tldextract==2.2.2
Twisted==19.10.0
uritemplate==3.0.1
urllib3==1.24.3
w3lib==1.21.0
warcio==1.7.1
Werkzeug==0.16.0
zope.interface==4.7.1
